Renamed CNN.py to NN.py, added models and training data
- Implemented classify button in GUI
- GUI now shows output layer of NN
- added a 93% accuracy model and MNIST training data
